updated server/ with routes/api folder. in this folder i provided 2 index.js files as well as a manager/technician-routes
